temporarily write some duplicate code
